# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestor Recursos API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api/v1"
        }
    ],
    "paths": {
...
        "/tag?searchNombre=''&sortNombre=''": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "List Tags",
                "description": "List Tags",
                "operationId": "getTags",
                "parameters": [
                    {
                        "name": "searchNombre",
                        "in": "query",
                        "description": "Filter Tag By Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortNombre",
                        "in": "query",
                        "description": "Sorting Name by desc or asc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "identificador": {
                                                        "type": "integer",
                                                        "example": 67
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "DESARROLLO WEB"
                                                    },
                                                    "estilos": {
                                                        "type": "string",
                                                        "example": "bg-pink-400 text-white"
                                                    },
                                                    "total": {
                                                        "type": "number",
                                                        "nullable": 5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag"
                                                },
                                                "currentPage": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "perPage": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "totalPages": {
                                                    "type": "number",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=4"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=2"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tag": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get All Tag to SelectTag",
                "description": "Get All Tag to SelectTag",
                "operationId": "getTagsToSelecTag",
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 67
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "DESARROLLO WEB"
                                                    },
                                                    "style": {
                                                        "type": "string",
                                                        "example": "bg-pink-400 text-white"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Save a Tag",
                "description": "Save a Tag",
                "operationId": "postTag",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mi Etiqueta"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Mi Etiqueta"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "identificador": {
                                                    "type": "integer",
                                                    "example": 67
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "DESARROLLO WEB"
                                                },
                                                "estilos": {
                                                    "type": "string",
                                                    "example": "bg-pink-400 text-white"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "nullable": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tag/{tag_id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get Tag",
                "description": "Get Tag",
                "operationId": "getTag",
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 202
                                        },
                                        "data": {
                                            "properties": {
                                                "identificador": {
                                                    "type": "integer",
                                                    "example": 67
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "DESARROLLO WEB"
                                                },
                                                "estilos": {
                                                    "type": "string",
                                                    "example": "bg-pink-400 text-white"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "nullable": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Update a Tag",
                "description": "Update a Tag",
                "operationId": "putTag",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mi Etiqueta"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Mi Etiqueta"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 202
                                        },
                                        "data": {
                                            "properties": {
                                                "identificador": {
                                                    "type": "integer",
                                                    "example": 67
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "DESARROLLO WEB"
                                                },
                                                "estilos": {
                                                    "type": "string",
                                                    "example": "bg-pink-400 text-white"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "nullable": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tag/{recourse_id}": {
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Delete a Tag",
                "description": "Delete a Tag",
                "operationId": "deleteTag",
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 202
                                        },
                                        "data": {
                                            "properties": {
                                                "identificador": {
                                                    "type": "integer",
                                                    "example": 67
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "DESARROLLO WEB"
                                                },
                                                "estilos": {
                                                    "type": "string",
                                                    "example": "bg-pink-400 text-white"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "nullable": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}